Added copy od released circuit

+++ arduino/projects2/barobot_carret2/source/barobot_carret2_main.h
@@ -10,7 +10,7 @@ uint8_t GetTemp();
 //-------------------------------------------------------------------
  
 // Put yout declarations here
- unsigned int decodeInt(String input, int odetnij );
+ unsigned int decodeInt(String input, byte odetnij );
 //-------------------------------------------------------------------
  
 //===================================================================
